+++ modules/core/src/test/java/org/apache/ignite/spi/communication/GridTestMessage.java
@@ -29,7 +29,7 @@ import org.apache.ignite.plugin.extensions.communication.MessageWriter;
  */
 public class GridTestMessage implements Message {
     /** */
-    public static final byte DIRECT_TYPE = (byte)200;
+    public static final short DIRECT_TYPE = 200;
 
     /** */
     private UUID srcNodeId;
